Move event dispatching to separate package
