+++ src/android/java/com/pspdfkit/cordova/action/SaveDocumentAction.java
@@ -1,12 +1,14 @@
     package com.pspdfkit.cordova.action;
 
-import com.pspdfkit.cordova.EventDispatcher;
+import com.pspdfkit.cordova.CordovaPdfActivity;
 import com.pspdfkit.cordova.PSPDFKitCordovaPlugin;
 
 import org.apache.cordova.CallbackContext;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
+
+import java.io.IOException;
 
 import androidx.annotation.NonNull;
 
@@ -15,8 +17,8 @@ public class SaveDocumentAction extends BasicAction {
         super(name, plugin);
     }
 
-    @Override protected void execAction(JSONArray args, CallbackContext callbackContext) throws JSONException {
-        final boolean wasModified = EventDispatcher.getInstance().notifySaveDocument();
+    @Override protected void execAction(JSONArray args, CallbackContext callbackContext) throws JSONException, IOException {
+        final boolean wasModified = CordovaPdfActivity.getCurrentActivity().saveDocument();
         final JSONObject response = new JSONObject();
         response.put("wasModified", wasModified);
         callbackContext.success(response);
